add authorization module

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.RemoveDish
@@ -9,7 +11,8 @@ namespace Restaurants.Application.Dishes.Commands.RemoveDish
     internal class DeleteDishesForRestaurantCommandHandler(
         ILogger<DeleteDishesForRestaurantCommand> logger, 
         IRestaurantsRepository restaurantsRepository,
-        IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+        IDishesRepository dishesRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
     {
         public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -17,6 +20,9 @@ namespace Restaurants.Application.Dishes.Commands.RemoveDish
 
             var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId, cancellationToken)
                 ?? throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+                throw new ForbidException();
 
             await dishesRepository.Delete(restaurant.Dishes, cancellationToken);
         }
